pyproject.toml
[project]
name = "newsnowcon"
version = "0.1.4"
description = "Power up your crews with newsnowcon"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.203.1",
    "snowflake-connector-python>=3.6.0"
]

[tool.crewai]
type = "tool"
